distributed code in differnet modules
